+++ src/Ext/WarheadType/Detonate.cpp
@@ -215,6 +215,9 @@ void WarheadTypeExt::ExtData::DetonateOnOneUnit(HouseClass* pHouse, TechnoClass*
 	if (this->BuildingSell || this->BuildingUndeploy)
 		this->ApplyBuildingUndeploy(pTarget);
 
+	if (this->Taunt)
+		this->ApplyTaunt(pHouse, pTarget, pOwner);
+
 #ifdef LOCO_TEST_WARHEADS
 	if (this->InflictLocomotor)
 		this->ApplyLocomotorInfliction(pTarget);
@@ -607,6 +610,46 @@ void WarheadTypeExt::ExtData::ApplyAttachEffects(TechnoClass* pTarget, HouseClas
 	AttachEffectClass::Attach(pTarget, pInvokerHouse, pInvoker, this->OwnerObject(), info);
 	AttachEffectClass::Detach(pTarget, info);
 	AttachEffectClass::DetachByGroups(pTarget, info);
+}
+
+void WarheadTypeExt::ExtData::ApplyTaunt(HouseClass* pHouse, TechnoClass* pTarget, TechnoClass* pOwner)
+{
+	if (!this->Taunt_AffectsBerserk && pTarget->Berzerk)
+		return;
+
+	auto const pTargetExt = TechnoExt::ExtMap.Find(pTarget);
+
+	if (pTargetExt->TypeExtData->ImmuneToTaunt)
+		return;
+
+	auto const pShield = pTargetExt->Shield.get();
+
+	if (pShield && pShield->IsActive() && pShield->GetType()->ImmuneToTaunt)
+		return;
+
+	if (!TechnoExt::IsActive(pTarget))
+		return;
+
+	if (!this->Taunt_AffectsControlledAllies && pTarget->IsMindControlled() && pHouse && pHouse->IsAlliedWith(pTarget->GetOriginalOwner()))
+		return;
+
+	if (pHouse && !EnumFunctions::CanTargetHouse(this->Crit_AffectsHouses, pHouse, pTarget->Owner))
+		return;
+
+	// clean current missions first to avoid failure
+	const auto pAircraft = abstract_cast<AircraftClass*, true>(pTarget);
+	const bool commonAircraft = pAircraft && !pAircraft->Airstrike && !pAircraft->Spawned;
+	TechnoExt::PreProcessStopCommand(pTarget, pAircraft, commonAircraft);
+
+	if (this->Taunt_BreakMission || !pOwner || pTarget->SelectWeapon(pOwner) < 0)
+	{
+		TechnoExt::ProcessStopCommand(pTarget, pAircraft, commonAircraft);
+	}
+	else
+	{
+		pTarget->QueueMission(Mission::Attack, true);
+		pTarget->SetTarget(pOwner);
+	}
 }
 
 double WarheadTypeExt::ExtData::GetCritChance(TechnoClass* pFirer) const